% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_gx1.R
\name{fn_GX1_CRD_BLUEs}
\alias{fn_GX1_CRD_BLUEs}
\title{Fit linear models to extract the best linear unbiased estimators for the genotype values,
assuming a completely randomised design (CRD) in a single environment.
We use this design when we do not expect heterogeneity in the entire trial area, e.g. small trial and highly-controlled conditions.}
\usage{
fn_GX1_CRD_BLUEs(df, control_id, trait = "y", id = "gen", verbose = FALSE)
}
\arguments{
\item{df}{data frame containing the model variables}

\item{control_id}{name of the control genotype you wish to set as the the intercept from which all other entries where compared against}

\item{trait}{name of the continuous numeric response variable in the data frame}

\item{id}{name of the entry field (main explanatory variable) in the data frame}

\item{verbose}{show model fitting messages?}
}
\value{
df_effects: data frame trial-estimated breeding values (fields: id, BLUEs)
loglik: log-likelihood of the best-fitting model
AIC:  Akaike information criterion (prediction error estimator) of the best-fitting model
BIC: Bayesian information criterion (another prediction error estimator) of the best-fitting model
algorithm: model fitting algorithm used, i.e. Henderson's (mmec) or Newton-Raphson-transformations (mmer)
model: Specification of the best-fitting linear model
V: Variance-covariance component of the fixed effects
}
\description{
Fit linear models to extract the best linear unbiased estimators for the genotype values,
assuming a completely randomised design (CRD) in a single environment.
We use this design when we do not expect heterogeneity in the entire trial area, e.g. small trial and highly-controlled conditions.
}
\examples{
df = fn_simulate_gx1(design="crd")
out = fn_GX1_CRD_BLUEs(df=df, control_id=unique(df$gen)[2], trait="y", id="gen", verbose=TRUE)
}
