% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_gx1.R
\name{fn_GX1_CRD_BLUPs}
\alias{fn_GX1_CRD_BLUPs}
\title{Fit mixed models to extract the best linear unbiased predictors for the genotype values,
assuming a completely randomised design (CRD) in a single environment.
We use this design when we do not expect heterogeneity in the entire trial area, e.g. small trial and highly-controlled conditions.}
\usage{
fn_GX1_CRD_BLUPs(df, trait = "y", id = "gen", verbose = FALSE)
}
\arguments{
\item{df}{data frame containing the model variables}

\item{trait}{name of the continuous numeric response variable in the data frame}

\item{id}{name of the entry field (main explanatory variable) in the data frame}

\item{verbose}{show model fitting messages?}
}
\value{
df_effects: data frame trial-estimated breeding values (fields: id, BLUPs)
loglik: log-likelihood of the best-fitting model
AIC:  Akaike information criterion (prediction error estimator) of the best-fitting model
BIC: Bayesian information criterion (another prediction error estimator) of the best-fitting model
algorithm: model fitting algorithm used, i.e. Henderson's (mmec) or Newton-Raphson-transformations (mmer)
model: Specification of the best-fitting linear model
V: Variance-covariance component of the random effects
}
\description{
Fit mixed models to extract the best linear unbiased predictors for the genotype values,
assuming a completely randomised design (CRD) in a single environment.
We use this design when we do not expect heterogeneity in the entire trial area, e.g. small trial and highly-controlled conditions.
}
\examples{
df = fn_simulate_gx1(design="crd")
out = fn_GX1_CRD_BLUPs(df=df, trait="y", id="gen", verbose=TRUE)
}
