% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_gx1.R
\name{fn_GX1_SPAT_BLUEs}
\alias{fn_GX1_SPAT_BLUEs}
\title{Fit mixed models to extract the best linear unbiased estimators for the genotype values,
assuming a spatially (SPAT) explicit design with row-by-column coordinates in a single environment.
We use this design when we expect significant heterogeneity in the entire trial area, e.g. field trials with known nutrient gradients along one direction and moisture gradient along an orthogonal direction.}
\usage{
fn_GX1_SPAT_BLUEs(
  df,
  control_id,
  trait = "y",
  id = "gen",
  row = "row",
  col = "col",
  verbose = FALSE
)
}
\arguments{
\item{df}{data frame containing the model variables}

\item{control_id}{name of the control genotype you wish to set as the the intercept from which all other entries where compared against}

\item{trait}{name of the continuous numeric response variable in the data frame}

\item{id}{name of the entry field (main explanatory variable) in the data frame}

\item{row}{name of the row field in the data frame}

\item{col}{name of the column field in the data frame}

\item{verbose}{show model fitting messages?}
}
\value{
df_effects: data frame trial-estimated breeding values (fields: id, BLUEs)
loglik: log-likelihood of the best-fitting model
AIC:  Akaike information criterion (prediction error estimator) of the best-fitting model
BIC: Bayesian information criterion (another prediction error estimator) of the best-fitting model
algorithm: model fitting algorithm used, i.e. Henderson's (mmec) or Newton-Raphson-transformations (mmer)
model: Specification of the best-fitting linear model
V: Variance-covariance component of the fixed effects
}
\description{
Fit mixed models to extract the best linear unbiased estimators for the genotype values,
assuming a spatially (SPAT) explicit design with row-by-column coordinates in a single environment.
We use this design when we expect significant heterogeneity in the entire trial area, e.g. field trials with known nutrient gradients along one direction and moisture gradient along an orthogonal direction.
}
\examples{
df = fn_simulate_gx1(design="spat")
out = fn_GX1_SPAT_BLUEs(df=df, control_id=unique(df$gen)[4], trait="y", id="gen", 
     row="row", col="col", verbose=TRUE)
}
