% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_gxe.R
\name{fn_GXE_CRD_BLUPs}
\alias{fn_GXE_CRD_BLUPs}
\title{Fit a GXE model to extract the best linear unbiased predictors for the genotype values,
assuming a completely randomised design (CRD) in multiple environments.
We use this design if per environment, we do not expect heterogeneity in each trial area, e.g. small trial and highly-controlled conditions.}
\usage{
fn_GXE_CRD_BLUPs(
  df,
  trait = "y",
  id = "gen",
  env = "env",
  tolParInv = 0.01,
  skip_algo_based_on_dimensions = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{data frame containing the model variables}

\item{trait}{name of the continuous numeric response variable in the data frame}

\item{id}{name of the entry field in the data frame}

\item{env}{name of the environment field in the data frame}

\item{tolParInv}{diagonal loading of the variance-covariance matrix prior to inversion}

\item{skip_algo_based_on_dimensions}{if there are potentially more effects to be estimated than the number of observations then skip Henderson's fitting in favour of the Newton-Raphson transformations, and vice-versa}

\item{verbose}{show model fitting messages?}
}
\value{
df_effects: data frame trial-estimated breeding values per environment (fields: id, env, BLUPs)
loglik: log-likelihood of the best-fitting model
AIC:  Akaike information criterion (prediction error estimator) of the best-fitting model
BIC: Bayesian information criterion (another prediction error estimator) of the best-fitting model
algorithm: model fitting algorithm used, i.e. Henderson's (mmec) or Newton-Raphson-transformations (mmer)
model: Specification of the best-fitting linear model
V: Variance-covariance component of the random effects
}
\description{
Fit a GXE model to extract the best linear unbiased predictors for the genotype values,
assuming a completely randomised design (CRD) in multiple environments.
We use this design if per environment, we do not expect heterogeneity in each trial area, e.g. small trial and highly-controlled conditions.
}
\examples{
df = fn_simulate_gx1(design="crd")
out = fn_GXE_CRD_BLUPs(df=df, trait="y", id="gen", env="env", tolParInv=0.01, 
     skip_algo_based_on_dimensions=TRUE, verbose=TRUE)
}
