% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fn_simulate_gx1}
\alias{fn_simulate_gx1}
\title{Wrapper function for the simulation of single-environment trial data}
\usage{
fn_simulate_gx1(
  design = c("crd", "rbd", "spat"),
  n = 100,
  l = 1000,
  ploidy = 42,
  n_alleles = 2,
  pheno_reps = 3
)
}
\arguments{
\item{design}{experimental design, i.e. "crd", "rbd", and "spat" for completely randomised design, randomised complete/incomplete block design and spatial designs, respectively}

\item{n}{number of samples (individuals or pools)}

\item{l}{number of loci}

\item{ploidy}{ploidy level of individual samples or the number of individuals multiplied by their ploidy to simulate pools}

\item{n_alleles}{number of alleles per locus}

\item{pheno_reps}{number of phenotype replications}
}
\value{
data frame of simulated trial data
}
\description{
Wrapper function for the simulation of single-environment trial data
}
